mdx_to_skeleton: 모듈 분리 및 리팩토링
## Description
- `mdx_to_skeleton.py`에서 코드를 분리하여 3개의 독립 모듈로 리팩토링
- `skeleton_common.py`: 공통 유틸리티 함수 (언어 코드 추출, 경로 변환 등)
- `skeleton_compare.py`: `--compare` 기능 관련 코드 분리
- `skeleton_diff.py`: skeleton diff 비교 및 재귀 처리 기능 분리
- 코드 가독성 및 유지보수성 개선
