+++ confluence-mdx/bin/mdx_to_skeleton.py
@@ -12,16 +12,17 @@ Usage:
 
 import argparse
 import re
-import subprocess
 import sys
 from pathlib import Path
 from typing import List, Tuple, Optional
 
-# Global diff counter for tracking the number of diffs found
-_diff_count: int = 0
-_match_count: int = 0  # Counter for matching skeleton diffs
-_max_diff: Optional[int] = None  # Will be set to 5 (default) when --recursive is used
-_exclude_patterns: List[str] = ['/index.skel.mdx']  # Default exclude patterns
+# Import modules for recursive processing and comparison
+from skeleton_compare import compare_files
+from skeleton_diff import (
+    compare_with_korean_skel,
+    process_directories_recursive,
+    initialize_config,
+)
 
 
 class ProtectedSection:
@@ -151,260 +152,6 @@ def restore_protected_sections(text: str, sections: List[ProtectedSection]) -> s
     for section in sections:
         text = text.replace(section.placeholder, section.content)
     return text
-
-
-def extract_language_code(file_path: Path) -> Optional[str]:
-    """
-    Extract language code from the file path.
-    Assumes a relative path starting with target/{lang}/.
-    Checks for 'ko', 'en', 'ja' in the target/{lang}/ pattern.
-    Returns language code if found, None otherwise.
-    """
-    path_str = str(file_path)
-    path_lower = path_str.lower()
-
-    # Check for language codes in the target/{lang}/ pattern at the start
-    for lang in ['ko', 'en', 'ja']:
-        pattern = r'^target[/\\]' + lang + r'[/\\]'
-        if re.match(pattern, path_lower, re.IGNORECASE):
-            return lang.lower()
-
-    return None
-
-
-def get_korean_equivalent_path(file_path: Path) -> Optional[Path]:
-    """
-    Gets the Korean equivalent path by replacing the language code in the path.
-    Assumes a relative path starting with target/{lang}/.
-    If file_path is target/en/file.mdx, returns target/ko/file.mdx.
-    If file_path is target/ja/file.mdx, returns target/ko/file.mdx.
-    If no language code found, returns None.
-    """
-    path_str = str(file_path)
-    path_lower = path_str.lower()
-
-    # Try to replace language codes in the target/{lang}/ pattern
-    for lang in ['en', 'ja']:
-        # Match the target/{lang}/ pattern at the start
-        pattern = r'^target[/\\]' + lang + r'[/\\]'
-        if re.match(pattern, path_lower, re.IGNORECASE):
-            # Replace target/{lang}/ with target/ko/
-            new_path_str = re.sub(pattern, 'target/ko/', path_str, flags=re.IGNORECASE, count=1)
-            return Path(new_path_str)
-
-    return None
-
-
-def get_path_without_lang_dir(file_path: Path) -> Optional[str]:
-    """
-    Extracts a path without the target/{lang} prefix.
-    Assumes a relative path starting with target/{lang}/.
-    For example: target/en/some/path/file.skel.mdx -> /some/path/file.skel.mdx
-    Returns None if target/{lang} pattern is not found.
-    """
-    path_str = str(file_path)
-    path_lower = path_str.lower()
-
-    # Check for the target/{lang}/ pattern at the start (relative path only)
-    for lang in ['ko', 'en', 'ja']:
-        pattern = r'^target[/\\]' + lang + r'[/\\]'
-        match = re.match(pattern, path_lower, re.IGNORECASE)
-        if match:
-            end = match.end()
-            relative_path = path_str[end:]
-            # Ensure it starts with /
-            if not relative_path.startswith('/'):
-                relative_path = '/' + relative_path
-            return relative_path
-
-    return None
-
-
-def get_original_mdx_path(skel_path: Path) -> Optional[Path]:
-    """
-    Gets the original .mdx file path from the .skel.mdx path.
-    For example: target/ko/file.skel.mdx -> target/ko/file.mdx
-    """
-    if not skel_path.name.endswith('.skel.mdx'):
-        return None
-    return skel_path.parent / skel_path.name.replace('.skel.mdx', '.mdx')
-
-
-def format_diff_with_original_content(
-        diff_output: str,
-        left_skel_path: Path,
-        right_skel_path: Path
-) -> str:
-    """
-    Formats diff output by replacing .skel.mdx file paths with .mdx paths
-    and replacing diff content lines with original .mdx file content.
-
-    Converts skeleton diff (with _TEXT_ placeholders) to original content diff.
-    """
-    # Get original .mdx file paths
-    left_mdx_path = get_original_mdx_path(left_skel_path)
-    right_mdx_path = get_original_mdx_path(right_skel_path)
-
-    if left_mdx_path is None or right_mdx_path is None:
-        return diff_output
-
-    # Read original .mdx files
-    try:
-        left_lines = left_mdx_path.read_text(encoding='utf-8').split('\n')
-        right_lines = right_mdx_path.read_text(encoding='utf-8').split('\n')
-    except FileNotFoundError:
-        return diff_output
-
-    # Unified format chunk header pattern: @@ -start,count +start,count @@
-    chunk_header_pattern = re.compile(r'^@@ -(\d+)(?:,(\d+))? \+(\d+)(?:,(\d+))? @@')
-
-    result_lines = []
-    lines = diff_output.split('\n')
-    left_line_num = None
-    right_line_num = None
-
-    for line in lines:
-        # Replace file paths in headers
-        if line.startswith('--- '):
-            result_lines.append(line.replace(str(left_skel_path), str(left_mdx_path)))
-            continue
-        elif line.startswith('+++ '):
-            result_lines.append(line.replace(str(right_skel_path), str(right_mdx_path)))
-            continue
-
-        # Process chunk header: initialize line number tracking
-        match = chunk_header_pattern.match(line)
-        if match:
-            result_lines.append(line)
-            left_line_num = int(match.group(1))
-            right_line_num = int(match.group(3))
-            continue
-
-        # Process content lines within a chunk
-        if left_line_num is not None and right_line_num is not None:
-            if line.startswith('-'):
-                # Deleted line from left file
-                if 0 < left_line_num <= len(left_lines):
-                    result_lines.append('-' + left_lines[left_line_num - 1])
-                else:
-                    result_lines.append(line)
-                left_line_num += 1
-            elif line.startswith('+'):
-                # Added line to right file
-                if 0 < right_line_num <= len(right_lines):
-                    result_lines.append('+' + right_lines[right_line_num - 1])
-                else:
-                    result_lines.append(line)
-                right_line_num += 1
-            elif line.startswith(' '):
-                # Context line (common to both files)
-                if 0 < left_line_num <= len(left_lines):
-                    result_lines.append(' ' + left_lines[left_line_num - 1])
-                else:
-                    result_lines.append(line)
-                left_line_num += 1
-                right_line_num += 1
-            else:
-                # Empty line or end of chunk - keep as is
-                result_lines.append(line)
-        else:
-            # Outside chunk - keep as is
-            result_lines.append(line)
-
-    return '\n'.join(result_lines)
-
-
-def compare_with_korean_skel(current_skel_path: Path) -> Tuple[bool, Optional[str]]:
-    """
-    Compare current .skel.mdx file with Korean equivalent if it exists.
-    If current file is not Korean, find Korean equivalent and run diff.
-
-    Returns:
-        Tuple of (should_continue, comparison_result)
-        should_continue: True if it should continue processing, False if max_diff is reached and should stop
-        comparison_result: 'matched' if files are identical, 'unmatched' if different, None if not compared
-    """
-    global _diff_count, _match_count, _max_diff, _exclude_patterns
-
-    current_lang = extract_language_code(current_skel_path)
-
-    # If current file is Korean, no need to compare
-    if current_lang == 'ko':
-        return True, None
-
-    # Check if file path matches exclude patterns
-    relative_path = get_path_without_lang_dir(current_skel_path)
-    if relative_path and relative_path in _exclude_patterns:
-        return True, None
-
-    # Check if max_diff is set and already reached
-    if _max_diff is not None:
-        if _diff_count >= _max_diff:
-            return False, None
-
-    # Get Korean equivalent path
-    korean_skel_path = get_korean_equivalent_path(current_skel_path)
-
-    if korean_skel_path is None:
-        return True, None
-
-    # Check if Korean .skel.mdx file exists
-    if not korean_skel_path.exists():
-        return True, None
-
-    # Run diff command
-    try:
-        # Build diff command with unified format (-U 2 for 2 lines of context)
-        diff_cmd = ['diff', '-U', '2', str(korean_skel_path), str(current_skel_path)]
-
-        # Print command with "+ " prefix
-        print(f"+ {' '.join(diff_cmd)}")
-
-        # Run diff and capture output
-        result = subprocess.run(
-            diff_cmd,
-            capture_output=True,
-            text=True,
-            check=False  # Don't raise exception on non-zero exit
-        )
-
-        # Check if files are different (exit code 1 means differences found)
-        # Exit code 0 means files are identical
-        if result.returncode == 1:
-            # Files are different, increment diff count
-            _diff_count += 1
-
-            # Print diff output
-            if result.stdout:
-                print(result.stdout, end='')
-
-            # Print original .mdx file diff
-            original_diff = format_diff_with_original_content(
-                result.stdout,
-                korean_skel_path,
-                current_skel_path
-            )
-            if original_diff:
-                print(original_diff, end='')
-
-            # Check if max_diff reached
-            if _max_diff is not None:
-                if _diff_count >= _max_diff:
-                    return False, 'unmatched'
-            return True, 'unmatched'
-        elif result.returncode == 0:
-            # Files are identical, increment match count
-            _match_count += 1
-            return True, 'matched'
-
-        # Print stderr if any (errors)
-        if result.stderr:
-            print(result.stderr, end='', file=sys.stderr)
-
-    except Exception as e:
-        print(f"Error running diff: {e}", file=sys.stderr)
-
-    return True, None
 
 
 def process_text_line(line: str) -> str:
@@ -962,211 +709,6 @@ def convert_mdx_to_skeleton(input_path: Path) -> Tuple[Path, Optional[str]]:
     return output_path, comparison_result
 
 
-def process_directory(directory: Path, recursive: bool = False) -> Tuple[int, int, int, int, int, List[str]]:
-    """
-    Process all .mdx files in a directory.
-    Returns tuple of (success_count, error_count, matched_count, unmatched_count, not_compared_count, not_compared_files).
-    """
-    if not directory.exists():
-        raise FileNotFoundError(f"Directory not found: {directory}")
-
-    if not directory.is_dir():
-        raise ValueError(f"Path is not a directory: {directory}")
-
-    success_count = 0
-    error_count = 0
-    matched_count = 0
-    unmatched_count = 0
-    not_compared_count = 0
-    not_compared_files = []  # Track files that were not compared
-
-    # Find all .mdx files
-    if recursive:
-        mdx_files = list(directory.rglob('*.mdx'))
-    else:
-        mdx_files = list(directory.glob('*.mdx'))
-
-    # Filter out .skel.mdx files
-    mdx_files = [f for f in mdx_files if not f.name.endswith('.skel.mdx')]
-
-    if not mdx_files:
-        return 0, 0, 0, 0, 0, []
-
-    for mdx_file in mdx_files:
-        # Check if max_diff reached before processing next file
-        global _diff_count, _max_diff
-        if _max_diff is not None and _diff_count >= _max_diff:
-            break
-
-        try:
-            _, comparison_result = convert_mdx_to_skeleton(mdx_file)
-            success_count += 1
-            
-            # Count matched/unmatched/not_compared
-            if comparison_result == 'matched':
-                matched_count += 1
-            elif comparison_result == 'unmatched':
-                unmatched_count += 1
-            elif comparison_result is None:
-                # Comparison was not performed (e.g., Korean file, no Korean equivalent, etc.)
-                not_compared_count += 1
-                # Get relative path for display
-                try:
-                    rel_path = mdx_file.relative_to(directory)
-                    not_compared_files.append(str(rel_path))
-                except ValueError:
-                    not_compared_files.append(str(mdx_file))
-
-            # Check again after processing (compare_with_korean_skel may have incremented _diff_count)
-            if _max_diff is not None and _diff_count >= _max_diff:
-                break
-        except ValueError as e:
-            # Skip .skel.mdx files silently
-            if '.skel.mdx' in str(e):
-                continue
-            print(f"{mdx_file}: {e}", file=sys.stderr)
-            error_count += 1
-        except Exception as e:
-            print(f"{mdx_file}: {e}", file=sys.stderr)
-            error_count += 1
-
-    return success_count, error_count, matched_count, unmatched_count, not_compared_count, not_compared_files
-
-
-def get_mdx_files(directory: Path) -> set[str]:
-    """
-    Get all .mdx files in a directory (recursively), excluding .skel.mdx files.
-    Returns a set of relative paths (as strings) from the directory root.
-    """
-    if not directory.exists() or not directory.is_dir():
-        return set()
-
-    mdx_files = list(directory.rglob('*.mdx'))
-    # Filter out .skel.mdx files
-    mdx_files = [f for f in mdx_files if not f.name.endswith('.skel.mdx')]
-
-    # Convert to relative paths from directory root
-    relative_paths = set()
-    for mdx_file in mdx_files:
-        try:
-            rel_path = mdx_file.relative_to(directory)
-            relative_paths.add(str(rel_path))
-        except ValueError:
-            # If relative_to fails, use absolute path
-            relative_paths.add(str(mdx_file))
-
-    return relative_paths
-
-
-def compare_files(verbose: bool = False):
-    """
-    Compare .mdx files across target/en, target/ja, and target/ko directories.
-    Outputs comparison results to stdout.
-
-    Args:
-        verbose: If False, skip files that exist in all three languages.
-                 If True, output all files.
-    """
-    base_dir = Path('target')
-    dirs = {
-        'ko': base_dir / 'ko',
-        'en': base_dir / 'en',
-        'ja': base_dir / 'ja',
-    }
-
-    # Get file lists for each directory
-    file_sets = {}
-    for lang, dir_path in dirs.items():
-        file_sets[lang] = get_mdx_files(dir_path)
-
-    # Get all unique file paths (union of all three sets)
-    all_files = file_sets['ko'] | file_sets['en'] | file_sets['ja']
-
-    # Sort alphabetically
-    sorted_files = sorted(all_files)
-
-    # Output comparison results
-    for file_path in sorted_files:
-        # Check existence in each directory
-        ko_exists = file_path in file_sets['ko']
-        en_exists = file_path in file_sets['en']
-        ja_exists = file_path in file_sets['ja']
-
-        # Skip if all three languages exist and not verbose
-        if not verbose and ko_exists and en_exists and ja_exists:
-            continue
-
-        # Format output: /path/to/file.mdx ko en ja
-        ko_status = 'ko' if ko_exists else '-'
-        en_status = 'en' if en_exists else '-'
-        ja_status = 'ja' if ja_exists else '-'
-
-        # Ensure path starts with /
-        output_path = file_path if file_path.startswith('/') else f'/{file_path}'
-        print(f"{output_path} {ko_status} {en_status} {ja_status}")
-
-
-def process_directories_recursive(directories: List[Path]) -> int:
-    """
-    Process multiple directories recursively.
-    If directories list is empty, uses default directories (target/ko, target/ja, target/en).
-    Returns exit code (0 for success).
-    """
-    if len(directories) == 0:
-        # No directories specified, use defaults (Korean, Japanese, English order)
-        default_dirs = [
-            Path('target/ko'),
-            Path('target/ja'),
-            Path('target/en')
-        ]
-        directories = default_dirs
-
-    total_success = 0
-    total_errors = 0
-    total_matched = 0
-    total_unmatched = 0
-    total_not_compared = 0
-
-    for directory in directories:
-        # Check if max_diff reached before processing next directory
-        global _diff_count, _max_diff
-        if _max_diff is not None and _diff_count >= _max_diff:
-            break
-
-        if not directory.exists():
-            print(f"Warning: Directory not found: {directory}", file=sys.stderr)
-            continue
-        if not directory.is_dir():
-            print(f"Warning: Path is not a directory: {directory}", file=sys.stderr)
-            continue
-        success_count, error_count, matched_count, unmatched_count, not_compared_count, not_compared_files = process_directory(directory, recursive=True)
-        total_success += success_count
-        total_errors += error_count
-        total_matched += matched_count
-        total_unmatched += unmatched_count
-        total_not_compared += not_compared_count
-
-        # Print statistics for this directory
-        # Verify: converted = matched + unmatched + not_compared
-        print(f"{directory}: {success_count} converted, {error_count} errors, {matched_count} matched, {unmatched_count} unmatched, {not_compared_count} not_compared")
-        
-        # Print not_compared files if any
-        if not_compared_files:
-            print(f"  Not compared files ({len(not_compared_files)}):")
-            for file_path in sorted(not_compared_files):
-                print(f"    - {file_path}")
-
-        # Check again after processing directory
-        if _max_diff is not None and _diff_count >= _max_diff:
-            break
-
-    # Print overall summary statistics
-    if len(directories) > 1:
-        print(f"Total: {total_success} converted, {total_errors} errors, {total_matched} matched, {total_unmatched} unmatched, {total_not_compared} not_compared")
-
-    return 0
-
-
 def main():
     parser = argparse.ArgumentParser(
         description='Convert MDX file(s) to skeleton format by replacing text with _TEXT_'
@@ -1212,14 +754,10 @@ def main():
 
     args = parser.parse_args()
 
-    # Set global max_diff and exclude_patterns if recursive mode is used
-    global _max_diff, _diff_count, _match_count, _exclude_patterns
+    # Initialize config if recursive mode is used
     if args.recursive is not None:
-        _max_diff = args.max_diff
-        _diff_count = 0  # Reset counter
-        _match_count = 0  # Reset match counter
-        # Use exclude patterns from args, or default if empty
-        _exclude_patterns = args.exclude if args.exclude and len(args.exclude) > 0 else ['/index.skel.mdx']
+        exclude_patterns = args.exclude if args.exclude and len(args.exclude) > 0 else ['/index.skel.mdx']
+        initialize_config(args.max_diff, exclude_patterns)
 
     try:
         if args.compare:
@@ -1228,7 +766,7 @@ def main():
             return 0
         elif args.recursive is not None:
             # Recursive mode: process directories
-            return process_directories_recursive(args.recursive)
+            return process_directories_recursive(args.recursive, convert_mdx_to_skeleton)
         elif args.input_path:
             # Single file mode
             if args.input_path.is_dir():
